refactor: extract stimulus-functions to separate module
